Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 
 int main() {
-    double n;
+    float n;
     printf("");
-    scanf("%lf", &n);
-    printf("You entered: %.4lf\n", n);
+    scanf("%f", &n);
+    printf("You entered: %.4f\n", n);
     return 0;
 }
